consumer_dashboard/requirements.txt
dash
plotly
kafka-python
python-dotenv
pandas
